requirements.txt
click==6.7
Django==1.11.5
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-filter==1.0.4
django-rq==0.9.6
django-tenants==1.3.3
djangorestframework==3.6.4
gunicorn==19.7.1
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3.1
python-dotenv==0.7.1
pytz==2017.2
redis==2.10.6
rq==0.8.2
